api/requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
httpx==0.27.0
pydantic==2.11.9
python-dotenv==1.1.1
email-validator==2.1.1
twilio==8.10.0
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9